refactor: 알림 fetching 한곳에서 받아오기
